// quiz/src/commons/types/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
    DateTime: any;
    /** The `Upload` scalar type represents a file upload. */
    Upload: any;
};

export type IBoard = {
    __typename?: "Board";
    _id: Scalars["ID"];
    boardAddress?: Maybe<IBoardAddress>;
    contents: Scalars["String"];
    createdAt: Scalars["DateTime"];
    deletedAt?: Maybe<Scalars["DateTime"]>;
    dislikeCount: Scalars["Int"];
    images?: Maybe<Array<Scalars["String"]>>;
    likeCount: Scalars["Int"];
    title: Scalars["String"];
    updatedAt: Scalars["DateTime"];
    user?: Maybe<IUser>;
    writer?: Maybe<Scalars["String"]>;
    youtubeUrl?: Maybe<Scalars["String"]>;
};

export type IBoardAddress = {
    __typename?: "BoardAddress";
    _id: Scalars["ID"];
    address?: Maybe<Scalars["String"]>;
    addressDetail?: Maybe<Scalars["String"]>;
    createdAt: Scalars["DateTime"];
    deletedAt?: Maybe<Scalars["DateTime"]>;
    updatedAt: Scalars["DateTime"];
    zipcode?: Maybe<Scalars["String"]>;
};
// ...
export type IUser = {
    __typename?: "User";
    _id: Scalars["ID"];
    createdAt: Scalars["DateTime"];
    deletedAt?: Maybe<Scalars["DateTime"]>;
    email: Scalars["String"];
    name: Scalars["String"];
    picture?: Maybe<Scalars["String"]>;
    updatedAt: Scalars["DateTime"];
    userPoint?: Maybe<IUserPoint>;
};

export type IUserPoint = {
    __typename?: "UserPoint";
    _id: Scalars["ID"];
    amount: Scalars["Int"];
    createdAt: Scalars["DateTime"];
    deletedAt?: Maybe<Scalars["DateTime"]>;
    updatedAt: Scalars["DateTime"];
    user: IUser;
};